lambda/StableDiffusionModelFunction/requirements.txt
sagemaker
pillow
numpy
